pyproject.toml
[tool.poetry]
name = "youtube-analytics-project"
version = "0.1.0"
description = ""
authors = ["Alexandp L. <alexandp83@mail.ru>"]
readme = "README.md"
packages = [{include = "youtube_analytics_project"}]

[tool.poetry.dependencies]
python = "^3.11"
google-api-python-client = "^2.86.0"
isodate = "^0.6.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
